fix: refactor indexes and remove circular dependencies
