+++ packages/ui-components/.eslintrc.json
@@ -15,5 +15,9 @@
   ],
   "rules": {
       "@typescript-eslint/explicit-function-return-type": "off"
-  }
+  },
+  "env": {
+    "browser": true,
+    "node": true
+},
 }
