+++ packages/ui-components/src/Templates/defaultLayout/index.ts
