+++ packages/ui-components/src/components/Atoms/displays/index.tsx
@@ -1,1 +1,1 @@
-export { AddressDisplay } from "./Address";
+export * from "./AddressDisplay";
