+++ packages/ui-components/src/components/Molecules/buttons/index.tsx
@@ -1,6 +1,5 @@
 export * from "./Button";
 export * from "./BackLinkButton";
-export * from "./UploadButton";
 export * from "./AccountButton";
 export * from "./NavButton";
 export * from "./IconButton";
