+++ packages/ui-components/src/components/Organisms/index.ts
@@ -5,3 +5,4 @@ export * from './BottomNav'
 export * from './NavBar'
 export * from './SideNav'
 export * from './TopBar'
+export * from './UploadButton'
