+++ packages/ui-components/src/components/Organisms/steps/index.ts
@@ -1,1 +1,1 @@
-export { default as Steps } from "./Steps";
+export * from './Steps'
