+++ packages/ui-components/src/index.tsx
@@ -1,13 +1,6 @@
 import "./styles.css";
 
-export * from "./components/Molecules/cards";
-export * from "./components/Molecules/buttons";
-export * from "./components/Molecules/badges";
-export * from "./components/Molecules/inputs";
-export * from "./components/Organisms/modals";
-export * from "./components/Organisms/table";
-export * from "./components/Molecules/alerts";
-export * from "./Templates/layout";
-export * from "./components/Atoms/displays";
+export * from "./Templates";
+export * from "./components";
 
 export * as utils from "./utils";
