+++ packages/ui-components/tailwind.config.js
@@ -1,4 +1,5 @@
 /** @type {import('tailwindcss').Config} */
+
 module.exports = {
 	content: [
 		"./src/**/*.{js,ts,jsx,tsx}",
